changed unobvious processing of flags
